Add files via upload
This is version 1 of my worlde unscrambler. This takes all the letters you have left as well as the position of any letters you've guessed correctly and spews out a bunch of English words for you.

Warning to those that try to do this with few guesses as you'll end up with n^5 guesses where n is the number of letters you're yet to eliminate.